# Remove QueuePolicy locks